Set project to export into user package registry
When built, the project will store the current build
directory in an abritrarily named file inside the user's
cmake package registry.
On Windows systems, this is stored in the registry.
On Unix systems, this is probably stored in ~/.cmake/packages/

This is to facilitate the process of using this library in
another cmake-built project.