added arguments to createPackfile, endianness
createPackfile has two more arguments to strip path from file and
the delimeter to use for stripping the path (either forwardslash
or backslash).

creating and reading packfiles are now endian aware. Integer values
written to the packfile will be written as little endian. If the system
uses big endian integers, they will be converted from little endian when
reading from the packfile, and converted to little endian when writing
to the packfile.